updated ws from courier
